Production/PC2/requirements.txt
absl-py==0.11.0
appdirs==1.4.4
asn1crypto==0.24.0
astroid==2.1.0
asttokens==1.1.13
astunparse==1.6.3
attrs==20.3.0
Automat==20.2.0
automationhat==0.2.0
beautifulsoup4==4.7.1
blinker==1.4
blinkt==0.1.2
buttonshim==0.0.2
cachetools==4.1.1
Cap1xxx==0.1.3
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
cmake==3.18.4.post1
colorama==0.3.7
colorzero==1.1
constantly==15.1.0
cookies==2.2.1
cryptography==2.6.1
cssselect==1.1.0
cycler==0.10.0
Cython==0.29.21
distlib==0.3.1
docutils==0.14
drumhat==0.1.0
entrypoints==0.3
envirophat==1.0.0
ExplorerHAT==0.4.2
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.9
fourletterphat==0.1.0
gast==0.3.3
gdown==3.12.2
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
gpiozero==1.5.1
grpcio==1.33.2
h5py==2.10.0
html5lib==1.0.1
hyperlink==20.0.1
idna==2.10
importlib-metadata==2.0.0
imutils==0.5.3
incremental==17.5.0
isort==4.3.4
itemadapter==0.2.0
itemloaders==1.0.4
itsdangerous==1.1.0
jedi==0.13.2
Jinja2==2.11.2
jmespath==0.10.0
jsonpickle==1.4.1
Keras==2.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyring==17.1.1
keyrings.alt==3.1.1
kiwisolver==1.3.1
lazy-object-proxy==1.3.1
logilab-common==1.4.2
lxml==4.3.2
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.3
mccabe==0.6.1
microdotphat==0.2.1
mock==4.0.2
mote==0.0.4
motephat==0.0.3
mypy==0.670
mypy-extensions==0.4.1
numpy==1.19.4
oauthlib==3.1.0
olefile==0.46
opt-einsum==3.3.0
pantilthat==0.0.7
parsel==1.6.0
parso==0.3.1
pgzero==1.2
phatbeat==0.1.1
pianohat==0.1.0
piglow==1.2.5
pigpio==1.44
Pillow==8.0.1
Protego==0.1.16
protobuf==3.14.0
psutil==5.7.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.6.1
pycrypto==2.6.1
PyDispatcher==2.0.5
pygame==1.9.4.post1
Pygments==2.3.1
PyGObject==3.30.4
PyHamcrest==2.0.2
pyinotify==0.9.6
PyJWT==1.7.0
pylint==2.2.2
pymongo==3.11.0
pyOpenSSL==19.0.0
pyparsing==2.4.7
pyserial==3.4
PySocks==1.7.1
python-apt==1.8.4.1
python-dateutil==2.8.1
pyxdg==0.25
PyYAML==5.3.1
queuelib==1.5.0
rainbowhat==0.1.0
requests==2.25.0
requests-oauthlib==1.3.0
responses==0.9.0
roman==2.0.0
RPi.GPIO==0.7.0
rsa==4.6
RTIMULib==7.2.1
scipy==1.4.1
Scrapy==2.4.1
scrollphat==0.0.7
scrollphathd==1.2.1
SecretStorage==2.3.1
Send2Trash==1.5.0
sense-hat==2.2.0
service-identity==18.1.0
simplejson==3.16.0
six==1.15.0
skywriter==0.0.7
sn3218==1.2.7
soupsieve==1.8
spidev==3.4
ssh-import-id==5.7
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
thonny==3.3.0
touchphat==0.0.1
tqdm==4.53.0
Twisted==20.3.0
twython==3.7.0
typed-ast==1.3.1
unicornhathd==0.0.4
urllib3==1.26.2
virtualenv==20.1.0
w3lib==1.22.0
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0
zope.interface==5.2.0
